JVM: move PSI-based LambdaInfo to PsiInlineCodegen
